dom/html/test/test_bug651956.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=651956
-->
<head>
  <title>Test for Bug 651956</title>
  <script type="application/javascript" src="/MochiKit/packed.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=651956">Mozilla Bug 651956</a>
<p id="display"></p>
<div id="content">
  <input>
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 651956 **/

var input = document.getElementsByTagName('input')[0];

var gotInputEvent = false;

input.addEventListener("input", function() {
  input.removeEventListener("input", arguments.callee, false);
  gotInputEvent = true;
}, false);

input.addEventListener("focus", function() {
  input.removeEventListener("focus", arguments.callee, false);
  synthesizeKey("VK_ESCAPE", {});

  setTimeout(function() {
    ok(!gotInputEvent, "No input event should have been sent.");
    SimpleTest.finish();
  });
}, false);

SimpleTest.waitForExplicitFinish();
SimpleTest.waitForFocus(function() {
  input.focus();
});

</script>
</pre>
</body>
</html>
